+++ src/rules.ts
@@ -12,6 +12,7 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
+import './css/content-selector-to-slotted';
 import './css/deprecated-css-custom-property-syntax';
 import './css/deprecated-shadow-dom-selectors';